+++ src/main/java/ladder/model/generator/RowGenerator.java
@@ -1,6 +1,5 @@
 package ladder.model.generator;
 
-import ladder.model.Direction;
 import ladder.model.Row;
 
 import java.util.ArrayList;
@@ -17,21 +16,7 @@ public class RowGenerator {
             linkedData.add(generateBooleanValue());
         }
 
-        return Row.of(makeDirections(linkedData));
-    }
-
-    public static List<Direction> makeDirections(final List<Boolean> linkedData) {
-        List<Direction> row = new ArrayList<>();
-        Direction direction = Direction.first(linkedData.get(0));
-
-        row.add(direction);
-        for (int i = 1; i < linkedData.size(); i++) {
-            direction = direction.next(linkedData.get(i));
-            row.add(direction);
-        }
-        row.add(direction.last());
-
-        return row;
+        return Row.of(DirectionsGenerator.makeDirections(linkedData));
     }
 
     private static boolean generateBooleanValue() {
