+++ src/test/java/ladder/model/generator/RowGeneratorTest.java
@@ -13,7 +13,7 @@ public class RowGeneratorTest {
     void 연결정보로_true와_false가_주어졌을때_테스트() {
         Direction first = Direction.first(true);
         Direction next = first.next(false);
-        assertThat(RowGenerator.makeDirections(Arrays.asList(true, false)))
+        assertThat(DirectionsGenerator.makeDirections(Arrays.asList(true, false)))
                 .isEqualTo(Arrays.asList(first, next, next.last()));
     }
 }
